a bunch of reorganization stuff i forgot to commit
