--- include/Helpers.hpp
@@ -1,13 +1,0 @@
-#ifndef HELPERS_HPP
-#define HELPERS_HPP
-
-#include "Helpers/AutoRegion.hpp"
-#include "Helpers/HoldKey.hpp"
-#include "Helpers/KeySequence.hpp"
-#include "Helpers/MenuEntryHelpers.hpp"
-#include "Helpers/OSDManager.hpp"
-#include "Helpers/QuickMenu.hpp"
-#include "Helpers/Strings.hpp"
-#include "Helpers/Wrappers.hpp"
-
-#endif
