+++ include/Megamix/Types.hpp
@@ -3,7 +3,7 @@
 
 #include <array>
 
-#include "types.h"
+//#include "types.h"
 
 namespace Megamix {
     struct Tempo {
