+++ src/Megamix/HookFuncs/Misc.cpp
@@ -1,0 +1,18 @@
+#include <3ds.h>
+#include <CTRPluginFramework.hpp>
+
+#include "Megamix.hpp"
+
+namespace Megamix::Hooks {
+    u32 getRegionCTR() {
+        //TODO: handle JP region / JP langpack
+        if (region == Region::KR)
+            return Region::KR_CTR;
+        return region;
+    }
+
+    u32 getRegionMegamix() {
+        //TODO: handle JP region / JP langpack
+        return region;
+    }
+}
